Dont use OpenParrot directory when writing files created by game
Battle gear 4 tuned when connected to a NESYS server reads and writes files from the server related to the event race on boot. While it can write these files to .\OpenParrot just fine it cannot read them due to missing/incorrect function detours and causes the game to freeze trying to find the files from the game root instead of .\OpenParrot when entering an event race. This is just an easy way to fix it by removing redirection to openparrot and storing these files with the game exe instead (Where the game expects them).